+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Support of monitoring mempool transactions (`mempoolTxs` query function).
 * Support of government related queries such as `constitution` (to query the current constitution definition) and `proposals` (to fetch for proposals that are considered for ratification)
 * Support of mempool based caching, enabled by setting `mempoolCache` in provider's configuration.
+* Support of local transaction submission based caching (i.e., submitted transactions are considered to know for to be made available outputs & spent outputs), enabled by setting `localTxSubmissionCache` in provider's configuration.
 
 ## 0.11.1